backend/requirements.txt
aiofiles==25.1.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
fastapi-utilities==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
pdfplumber==0.11.9
pillow==12.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
